split lxMath.h into smaller logical parts
