+++ loch/lxMath.h
@@ -31,10 +31,6 @@
 
 // Standard libraries
 #ifndef LXDEPCHECK
-#include <cstddef>
-#include <list>
-#include <vector>
-#include <map>
 #endif  
 //LXDEPCHECK - standard libraries
 
@@ -89,52 +85,6 @@ struct lxVecLimits {
   lxVecLimits Rotate(double a, double i, lxVec c);
   lxVec Vertex(int v);
 };
-
-
-struct lxTriGeomPoint {
-  lxVec p, n;
-  operator lxVec & () {return this->p;}
-  const lxVec & operator = (const lxVec q) {this->p = q; return this->p;}
-};
-
-
-bool operator < ( const lxTriGeomPoint& p, const lxTriGeomPoint& q );
-
-
-struct lxTriGeom3Angle {
-  size_t v1, v2, v3;
-  lxTriGeom3Angle() : v1(0), v2(0), v3(0) {}
-  lxTriGeom3Angle(size_t vv1, size_t vv2, size_t vv3);
-};
-
-
-struct lxTriGeom {
-  size_t m_nPoints;
-  std::map<lxTriGeomPoint, size_t> m_points;
-  std::vector<lxTriGeomPoint> m_positions;
-  std::list<lxTriGeom3Angle> m_triangles;
-  std::vector<lxTriGeom3Angle> m_geometry;
-  void Clear();
-  void Append(lxTriGeom * src);
-  size_t InsertPoint(lxTriGeomPoint p);
-  bool Insert3Angle(lxTriGeomPoint p1, lxTriGeomPoint p2, lxTriGeomPoint p3);
-  bool Insert4Angle(lxTriGeomPoint p1, lxTriGeomPoint p2, lxTriGeomPoint p3, lxTriGeomPoint p4);
-  size_t GetNPoints();
-  lxTriGeomPoint GetPoint(size_t i);
-  size_t GetNTriangles();
-  lxTriGeom3Angle GetTriangle(size_t i);
-  lxTriGeom() : m_nPoints(0) {}
-};
-
-
-struct lxPlane {
-  lxVec m_point, m_normal;
-  double m_delta = {};
-  void Init(lxVec pt, lxVec norm);
-  double CalcPosition(lxVec pt);
-  bool CalcIntersection(lxVec fp, lxVec tp, lxVec & tgt);
-};
-
 
 #define lxVecXYZ(vec) (vec).x, (vec).y, (vec).z
 #define lxMax(a,b) ((a) > (b) ? (a) : (b))
